Use 32-bit integers for line numbers
This saves 16% time, 18% memory and 1% instruction counts from the
sample files in #293. These are large files, treated as textual diffs,
so reducing the storage per-line is significant. Files with structural
diffing, such as sample_files/slow_before.rs, are unchanged.

Fixes #293
Closes #297

+++ src/diff/sliders.rs
@@ -538,7 +538,7 @@ fn slide_to_next_node<'a>(
 
 /// Return the distance between two syntax nodes, as a tuple of number
 /// of lines and number of columns.
-fn distance_between(prev: &Syntax, next: &Syntax) -> (usize, u32) {
+fn distance_between(prev: &Syntax, next: &Syntax) -> (u32, u32) {
     let prev_pos = prev.last_line_span();
     let next_pos = next.first_line_span();
